+++ app/src/main/java/com/example/meetup/model/SearchResponseModel.kt
@@ -23,7 +23,7 @@ data class StoreDto(
     val name: String,
     val description: String,
     val koreanYn: Boolean,
-    val avgRate: Long,
+    val avgRate: Double,
     val minOrderAmount: Long,
     val images: List<String>,
 )
@@ -33,8 +33,8 @@ data class SearchFood(
     val name: String,
     val storeId: Long,
     val storeName: String,
-    val dollarPrice: Long,
-    val canadaPrice: Long,
+    val dollarPrice: Double,
+    val canadaPrice: Double,
     val image: String?,
-    val avgRate: Long,
+    val avgRate: Double,
 )
